# Graph int speedup