@snowleopard 2018-06-11
@nobrakal That's great, thank you :)

I think you should also add `toIntAdjacencyMap` here: 

https://github.com/snowleopard/alga/blob/master/src/Algebra/Graph/ToGraph.hs#L250