@snowleopard 2018-06-11
By the way, I think the reported performance boost is often not very accurate, because we neglect additive constant factors. What if instead of trying to find `k` that works best for across a set of measurements `x[i]` and `y[i]` so that `x[i] ~ y[i] * k`, we find a best linear fit? That is, find optimal `k` and `a` so that we have the best fit `x[i] ~ y[i] * k + a` and report `k`?

See https://en.wikipedia.org/wiki/Linear_regression