@nobrakal 2018-06-11
Mmhh you are right, I never searched more than the Mean time from Criterion, but it purpose a 'regressed' time, from an "ordinary least-squares regression model."  (from http://www.serpentine.com/criterion/tutorial.html#understanding-the-data-under-a-chart ).
I will switch the bench suite to use that time