@snowleopard 2018-06-11
@nobrakal I think you misunderstood me: I'm not saying you should change the way in which you use Criterion. It's an improvement to your code. Instead of computing `k` as average `x[i]/y[i]`, you should compute it by solving a linear regression, i.e. something like this:
```
k = sum (x[i] - avg(x))(y[i] - avg(y)) / sum ((x[i] - avg(x))^2)
```