@nobrakal 2018-06-12
If my implementation is right, there is:

## vertexList

### Mesh
<TABLE>
   <TR>
      <TH>
      </TH>
      <TH CLASS = "thinright">
         1
      </TH>
      <TH CLASS = "thinright">
         10
      </TH>
      <TH CLASS = "thinright">
         100
      </TH>
      <TH>
         1000
      </TH>
   </TR>
   <TR>
      <TH>
         Alga
      </TH>
      <TD CLASS = "thinright">
         46.17 ns
      </TD>
      <TD CLASS = "thinright">
         622.7 ns
      </TD>
      <TD CLASS = "thinright">
         10.79 μs
      </TD>
      <TD>
         177.3 μs
      </TD>
   </TR>
   <TR>
      <TH>
         AlgaOld
      </TH>
      <TD CLASS = "thinright">
         44.52 ns
      </TD>
      <TD CLASS = "thinright">
         4.672 μs
      </TD>
      <TD CLASS = "thinright">
         122.3 μs
      </TD>
      <TD>
         2.207 ms
      </TD>
   </TR>
</TABLE>

### Clique
<TABLE>
   <TR>
      <TH>
      </TH>
      <TH CLASS = "thinright">
         1
      </TH>
      <TH CLASS = "thinright">
         10
      </TH>
      <TH CLASS = "thinright">
         100
      </TH>
      <TH>
         1000
      </TH>
   </TR>
   <TR>
      <TH>
         Alga
      </TH>
      <TD CLASS = "thinright">
         46.23 ns
      </TD>
      <TD CLASS = "thinright">
         1.790 μs
      </TD>
      <TD CLASS = "thinright">
         359.5 μs
      </TD>
      <TD>
         84.12 ms
      </TD>
   </TR>
   <TR>
      <TH>
         AlgaOld
      </TH>
      <TD CLASS = "thinright">
         44.45 ns
      </TD>
      <TD CLASS = "thinright">
         16.28 μs
      </TD>
      <TD CLASS = "thinright">
         3.737 ms
      </TD>
      <TD>
         634.0 ms
      </TD>
   </TR>
</TABLE>


SUMMARY:

 * Alga was the fastest 6 times
There was 2 ex-aequo


ABSTRACT:
(Calculated using simple linear regression)

 * Alga was 7.53 times faster than AlgaOld

Does it seems more coherent to you ?