@snowleopard 2018-06-12
> Does it seems more coherent to you ?

Hmm, looks like now it's completely dominated by the last term: `7.53 ~ 634.0/84.12`. Maybe that's the right thing to do indeed -- the largest benchmark is indeed the most sensible indicator of speedup.