@snowleopard 2018-06-12
@nobrakal I think I'm going to merge this, but I'll probably do some follow-up changes. I think the current naming is not very consistent: we have `adjacencyList` but `toAdjacencyMap`. I guess we should rename the latter simply to `adjacencyMap`.