@nobrakal 2018-06-12
Yes indeed this is the right thing to do, we want to promote results for big benchs !